databank doet nu onDelete: cascade voor assignments

+++ dwengo_backend/services/teacherAssignmentService.ts
@@ -1,9 +1,15 @@
-import {Assignment, PrismaClient} from "@prisma/client";
+import {Assignment, PrismaClient, Role} from "@prisma/client";
+import {isAuthorized} from "./authorizationService";
 
 const prisma = new PrismaClient();
 
 export const createAssignmentForClass = async (teacherId: number, classId: number, learningPathId: number): Promise<Assignment> => {
-    const assignment = await prisma.assignment.create({
+
+    if (!await isAuthorized(teacherId, Role.TEACHER, classId)) {
+        throw new Error("The teacher is unauthorized to perform this action");
+    }
+
+    const assignment: Assignment = await prisma.assignment.create({
         data: {
             learningPathId,
         },
@@ -22,7 +28,12 @@ export const createAssignmentForClass = async (teacherId: number, classId: numbe
     return assignment;
 }
 
-export const getAssignmentsByClass = async(classId: number): Promise<Assignment[]> => {
+export const getAssignmentsByClass = async(classId: number, teacherId: number): Promise<Assignment[]> => {
+
+    if (!await isAuthorized(teacherId, Role.TEACHER)) {
+        throw new Error("The teacher is unauthorized to request the assignments");
+    }
+
     const classAssignments = await prisma.classAssignment.findMany({
         where: {classId},
         include: {assignment: true},
@@ -31,23 +42,25 @@ export const getAssignmentsByClass = async(classId: number): Promise<Assignment[
     return classAssignments.map(ca => ca.assignment);
 }
 
-export const updateAssignment = async(assignmentId: number, learningPathId: number): Promise<Assignment> => {
+export const updateAssignment = async(assignmentId: number, learningPathId: number, teacherId: number): Promise<Assignment> => {
+
+    if (!await isAuthorized(teacherId, Role.TEACHER)) {
+        throw new Error("The teacher is unauthorized to update the assignment");
+    }
+
     return prisma.assignment.update({
         where: {id: assignmentId},
         data: {learningPathId},
     });
 }
 
-export const deleteAssignment = async (assignmentId: number) => {
-    return prisma.$transaction([
-        prisma.classAssignment.deleteMany({
-            where: { assignmentId },
-        }),
-        prisma.teamAssignment.deleteMany({
-            where: { assignmentId },
-        }),
-        prisma.assignment.delete({
-            where: { id: assignmentId },
-        }),
-    ]);
+export const deleteAssignment = async (assignmentId: number, teacherId: number): Promise<Assignment> => {
+
+    if (!await isAuthorized(teacherId, Role.TEACHER)) {
+        throw new Error("The teacher is unauthorized to delete the assignment");
+    }
+
+    return prisma.assignment.delete({
+        where: { id: assignmentId },
+    });
 };
